requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & RAG
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
faiss-cpu==1.9.0.post1
numpy==1.26.3

# Document Parsing
pdfplumber==0.10.4
pandas==2.2.0
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
